Issue #14 Rename console export with debug
