+++ lib/index.js
@@ -1,6 +1,6 @@
-import * as console from './console'
+import * as debug from './debug'
 import * as db from './db'
 import * as validate from './validate'
 import * as microservice from './microservice'
 
-export {console, db, validate, microservice}
+export {debug, db, validate, microservice}
